feat: throw error when slurmNode does not match any kubeNodes
Presently there is no error thrown when a slurmNode does not match any
kubeNodes during pre-filter. This can be confusing when first deploying
Slurm-bridge, as users may not have configured node labels yet. The
introduction of this error case makes it easier for users to understand
why their jobs are not being scheduled by slurm-bridge.

Changelog: added - Error when slurmNode does not match any kubeNodes in
 preFilter stage