feat: create placeholder jobs in PostFilter
Instead of relying on PreFilter to create the slurm placeholder job,
use PostFilter instead. This allows slurm-bridge to use kube's filtered
list of nodes as input to the placeholder job so slurm will select nodes
based on the list of nodes kube has already vetted. Doing so enables
use of upstream plugins like TaintToleration and VolumeClaim so pods
that set taints or PVC are handled correctly.

Changelog: Changed - Create the Slurm placeholder job in PostFilter
 instead of PreFilter.