feat: pod controller should clean up ResourceClaims for terminal pods
The DRA controller does not remove ResourceClaims automatically as
other pods may eventually use or share a claim. Because Slurm is
tightly controlling which pods gain access to GRES resources, a
ResourceClaim should be removed once the pod has entered a terminal
phase.

Changelog: Changed - A ResourceClaim deletion to the pod controller
 for pods that are completed or being deleted.