fix: explicitly enable and disable scheduler plugins
This commit enables the correct scheduler plugins and disables
all others by default using multipoint. SlurmBridge is not included
via multipoint to control ordering of plugins.

This is not currently needed to fix a scheduling problem, rather it
is being made to avoid problems in the future of default plugins
changing the behavior of slurm-bridge unexpectedly.

Changelog: Changed - Set enabled and disabled scheduler plugins
 explicitly and disable all by default with multipoint.